# simpeg-go/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.StatusKepegawaian:
    properties:
      created_at:
        type: string
      id:
        type: integer
      is_aktif:
        type: boolean
      kode:
        type: string
      nama:
        type: string
      updated_at:
        type: string
    type: object
  domain.UnitKerja:
    properties:
      children:
        items:
          $ref: '#/definitions/domain.UnitKerja'
        type: array
      created_at:
        type: string
      id:
        type: integer
      is_aktif:
        type: boolean
      kode:
        type: string
      nama:
        type: string
      parent_id:
        type: integer
      singkatan:
        type: string
      tipe:
        type: string
      updated_at:
        type: string
    type: object
...
  http.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  http.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    name: SIMPEG UML
  description: Backend API Sistem Informasi Manajemen Kepegawaian Universitas Muhammadiyah
    Lampung (Gin + GORM).
  termsOfService: http://simpeg.uml.ac.id
  title: SIMPEG UML API
  version: "1.0"
paths:
...
  /status-kepegawaian:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.StatusKepegawaian'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Daftar status kepegawaian
      tags:
      - Master Data
    post:
      consumes:
      - application/json
      parameters:
      - description: Data status kepegawaian
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.StatusKepegawaian'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.StatusKepegawaian'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Tambah status kepegawaian
      tags:
      - Master Data
  /status-kepegawaian/{id}:
    delete:
      parameters:
      - description: ID status kepegawaian
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Hapus status kepegawaian
      tags:
      - Master Data
    get:
      parameters:
      - description: ID status kepegawaian
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.StatusKepegawaian'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Detail status kepegawaian
      tags:
      - Master Data
    put:
      consumes:
      - application/json
      parameters:
      - description: ID status kepegawaian
        in: path
        name: id
        required: true
        type: integer
      - description: Data status kepegawaian
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.StatusKepegawaian'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.StatusKepegawaian'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Ubah status kepegawaian
      tags:
      - Master Data
  /unit-kerja:
    get:
      description: Daftar unit kerja; gunakan ?tree=1 untuk hierarki unit kerja.
      parameters:
      - description: 1 untuk hierarki
        in: query
        name: tree
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.UnitKerja'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Daftar unit kerja
      tags:
      - Master Data
    post:
      consumes:
      - application/json
      parameters:
      - description: Data unit kerja (kode, nama, tipe)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.UnitKerja'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.UnitKerja'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Tambah unit kerja
      tags:
      - Master Data
  /unit-kerja/{id}:
    delete:
      parameters:
      - description: ID unit kerja
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Hapus unit kerja
      tags:
      - Master Data
    get:
      parameters:
      - description: ID unit kerja
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.UnitKerja'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Detail unit kerja
      tags:
      - Master Data
    put:
      consumes:
      - application/json
      parameters:
      - description: ID unit kerja
        in: path
        name: id
        required: true
        type: integer
      - description: Data unit kerja
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.UnitKerja'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.UnitKerja'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Ubah unit kerja
      tags:
      - Master Data
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"